+++ lib/rubocop/cop/lint/nested_method_definition.rb
@@ -26,6 +26,10 @@ module RuboCop
           (block (send _ {:instance_eval :class_eval :module_eval} ...) ...)
         PATTERN
 
+        def_node_matcher :class_or_module_new_call?, <<-PATTERN
+          (block (send (const nil {:Class :Module}) :new) ...)
+        PATTERN
+
         def on_method_def(node, _method_name, _args, _body)
           find_nested_defs(node) do |nested_def_node|
             add_offense(nested_def_node, :expression)
@@ -36,7 +40,7 @@ module RuboCop
           node.each_child_node do |child|
             if child.def_type? || child.defs_type?
               yield child
-            elsif !eval_call?(child)
+            elsif !(eval_call?(child) || class_or_module_new_call?(child))
               find_nested_defs(child, &block)
             end
           end