+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 * [#2691](https://github.com/bbatsov/rubocop/pull/2691): Do not register an offense in `Performance/TimesMap` for calling `map` or `collect` on a variable named `times`. ([@rrosenblum][])
 * [#2689](https://github.com/bbatsov/rubocop/pull/2689): Change `Performance/RedundantBlockCall` to respect parentheses usage. ([@rrosenblum][])
 * [#2694](https://github.com/bbatsov/rubocop/issues/2694): Fix caching when using a different JSON gem such as Oj. ([@stormbreakerbg][])
+* [#2707](https://github.com/bbatsov/rubocop/pull/2707): Change `Lint/NestedMethodDefinition` to respect `Class.new` and `Module.new`. ([@owst][])
 
 ### Changes
 
@@ -1911,3 +1912,4 @@
 [@mattparlane]: https://github.com/mattparlane
 [@drenmi]: https://github.com/drenmi
 [@stormbreakerbg]: https://github.com/stormbreakerbg
+[@owst]: https://github.com/owst